+++ services/powerbidedicated/mgmt/2017-10-01/powerbidedicated/CHANGELOG.md
@@ -1,2 +1,28 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Struct Changes
+
+#### Removed Struct Fields
+
+1. ErrorResponse.Code
+1. ErrorResponse.Message
+
+## Additive Changes
+
+### New Funcs
+
+1. DedicatedCapacityMutableProperties.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Structs
+
+1. ErrorResponseError
+
+#### New Struct Fields
+
+1. DedicatedCapacityMutableProperties.Mode
+1. DedicatedCapacityProperties.Mode
+1. ErrorResponse.Error